refactor: rename `state.ts` to `promise_state.ts`
