+++ mod.ts
@@ -2,10 +2,10 @@ export * from "./barrier.ts";
 export * from "./lock.ts";
 export * from "./mutex.ts";
 export * from "./notify.ts";
+export * from "./promise_state.ts";
 export * from "./queue.ts";
 export * from "./rw_lock.ts";
 export * from "./semaphore.ts";
 export * from "./stack.ts";
-export * from "./state.ts";
 export * from "./testutil.ts";
 export * from "./wait_group.ts";
