+++ queue_test.ts
@@ -3,7 +3,7 @@ import {
   assertEquals,
   assertRejects,
 } from "https://deno.land/std@0.211.0/assert/mod.ts";
-import { promiseState } from "./state.ts";
+import { promiseState } from "./promise_state.ts";
 import { Queue } from "./queue.ts";
 
 Deno.test("Queue", async (t) => {
